mysite/requirements.txt
asgiref==3.9.1
attrs==25.4.0
Django==5.2.6
django-debug-toolbar==6.0.0
django-filter==25.1
djangorestframework==3.14.0
docutils==0.22.2
drf-spectacular==0.28.0
flake8==7.3.0
flake8-docstrings==1.7.0
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
mccabe==0.7.0
pillow==11.3.0
pycodestyle==2.14.0
pydocstyle==6.3.0
pyflakes==3.4.0
pytz==2025.2
PyYAML==6.0.3
referencing==0.36.2
rpds-py==0.27.1
snowballstemmer==3.0.1
sqlparse==0.5.3
typing_extensions==4.15.0
uritemplate==4.2.0
